Significant refactoring after talking to Martyn
plotfunctions.py -> axesfunctions.py
plotfunctions3D.py -> axesfunctions3D.py
helperfunctions.py -> datafunctions.py
plot functions from workbench.plotting.functions.py -> plotfunctions.py
(colour map functionality remains in workbench.plotting.functions.py)

Also enabled window, waterfall and clearWindow functionality in plotSpectrum and plotBin
