+++ Framework/PythonInterface/mantid/plots/plotfunctions.py
@@ -5,760 +5,360 @@
 #     & Institut Laue - Langevin
 # SPDX - License - Identifier: GPL - 3.0 +
 #  This file is part of the mantid package
-#
-#
-from __future__ import (absolute_import, division, print_function)
+from __future__ import absolute_import
 
+# std imports
+import math
+import numpy as np
 import collections
-import matplotlib
-import matplotlib.collections as mcoll
-import matplotlib.colors
-import matplotlib.dates as mdates
-import matplotlib.image as mimage
-import numpy
-import sys
 
-import mantid.api
-import mantid.kernel
-import mantid.plots.modest_image
-from mantid.plots.helperfunctions import get_axes_labels, get_bins, get_data_uneven_flag, get_distribution, \
-    get_matrix_2d_ragged, get_matrix_2d_data, get_md_data1d, get_md_data2d_bin_bounds, \
-    get_md_data2d_bin_centers, get_normalization, get_sample_log, get_spectrum, get_uneven_data, \
-    get_wksp_index_dist_and_label, check_resample_to_regular_grid, get_indices, get_normalize_by_bin_width
-from mantid.plots.utility import MantidAxType
+# 3rd party imports
+from matplotlib.gridspec import GridSpec
+from matplotlib.legend import Legend
 
-# Used for initializing searches of max, min values
-_LARGEST, _SMALLEST = float(sys.maxsize), -sys.maxsize
+# local imports
+from mantid.api import AnalysisDataService, MatrixWorkspace
+from mantid.kernel import ConfigService
+from mantid.plots import datafunctions, MantidAxes
+from mantid.py3compat import is_text_string, string_types
+
+# -----------------------------------------------------------------------------
+# Constants
+# -----------------------------------------------------------------------------
+PROJECTION = 'mantid'
+
+MARKER_MAP = {'square': 's', 'plus (filled)': 'P', 'point': '.', 'tickdown': 3,
+              'triangle_right': '>', 'tickup': 2, 'hline': '_', 'vline': '|',
+              'pentagon': 'p', 'tri_left': '3', 'caretdown': 7,
+              'caretright (centered at base)': 9, 'tickright': 1,
+              'caretright': 5, 'caretleft': 4, 'tickleft': 0, 'tri_up': '2',
+              'circle': 'o', 'pixel': ',', 'caretleft (centered at base)': 8,
+              'diamond': 'D', 'star': '*', 'hexagon1': 'h', 'octagon': '8',
+              'hexagon2': 'H', 'tri_right': '4', 'x (filled)': 'X',
+              'thin_diamond': 'd', 'tri_down': '1', 'triangle_left': '<',
+              'plus': '+', 'triangle_down': 'v', 'triangle_up': '^', 'x': 'x',
+              'caretup': 6, 'caretup (centered at base)': 10,
+              'caretdown (centered at base)': 11, 'None': 'None'}
+
+# -----------------------------------------------------------------------------
+# Decorators
+# -----------------------------------------------------------------------------
+def manage_workspace_names(func):
+    """
+    A decorator to go around plotting functions.
+    This will retrieve workspaces from workspace names before
+    calling the plotting function
+    :param func: A plotting function
+    :return:
+    """
+
+    def inner_func(workspaces, *args, **kwargs):
+        workspaces = _validate_workspace_names(workspaces)
+        return func(workspaces, *args, **kwargs)
+
+    return inner_func
 
 
-# ================================================
-# Private 2D Helper functions
-# ================================================
+# -----------------------------------------------------------------------------
+# Public Methods
+# -----------------------------------------------------------------------------
+def figure_title(workspaces, fig_num):
+    """Create a default figure title from a single workspace, list of workspaces or
+    workspace names and a figure number. The name of the first workspace in the list
+    is concatenated with the figure number.
 
-def _setLabels1D(axes, workspace, indices=None, normalize_by_bin_width=True,
-                 axis=MantidAxType.SPECTRUM):
-    '''
-    helper function to automatically set axes labels for 1D plots
-    '''
-    labels = get_axes_labels(workspace, indices, normalize_by_bin_width)
-    # We assume that previous checking has ensured axis can only be 1 of 2 types
-    axes.set_xlabel(labels[1 if axis == MantidAxType.SPECTRUM else 2])
-    axes.set_ylabel(labels[0])
+    :param workspaces: A single workspace, list of workspaces or workspace name/list of workspace names
+    :param fig_num: An integer denoting the figure number
+    :return: A title for the figure
+    """
 
+    def wsname(w):
+        return w.name() if hasattr(w, 'name') else w
 
-def _setLabels2D(axes, workspace, indices=None, transpose=False,
-                 xscale=None, normalize_by_bin_width=True):
-    '''
-    helper function to automatically set axes labels for 2D plots
-    '''
-    labels = get_axes_labels(workspace, indices, normalize_by_bin_width)
-    if transpose:
-        axes.set_xlabel(labels[2])
-        axes.set_ylabel(labels[1])
+    if is_text_string(workspaces) or not isinstance(workspaces, collections.Sequence):
+        # assume a single workspace
+        first = workspaces
     else:
-        axes.set_xlabel(labels[1])
-        axes.set_ylabel(labels[2])
-    axes.set_title(labels[0])
-    if xscale is None and hasattr(workspace, 'isCommonLogBins') and workspace.isCommonLogBins():
-        axes.set_xscale('log')
-    elif xscale is not None:
-        axes.set_xscale(xscale)
+        assert len(workspaces) > 0
+        first = workspaces[0]
 
+    return wsname(first) + '-' + str(fig_num)
 
-def _get_data_for_plot(axes, workspace, kwargs, with_dy=False, with_dx=False):
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        (x, y, dy) = get_md_data1d(workspace, normalization, indices)
-        dx = None
-        axis = None
+@manage_workspace_names
+def plot(workspaces, spectrum_nums=None, wksp_indices=None, errors=False,
+         overplot=False, fig=None, plot_kwargs=None, ax_properties=None,
+         window_title=None, tiled=False, waterfall=False):
+    """
+    Create a figure with a single subplot and for each workspace/index add a
+    line plot to the new axes. show() is called before returning the figure instance. A legend
+    is added.
+
+    :param workspaces: A list of workspace handles or strings
+    :param spectrum_nums: A list of spectrum number identifiers (general start from 1)
+    :param wksp_indices: A list of workspace indexes (starts from 0)
+    :param errors: If true then error bars are added for each plot
+    :param overplot: If true then overplot over the current figure if one exists. If an axis object the overplotting
+    will be done on the axis passed in
+    :param fig: If not None then use this Figure object to plot
+    :param plot_kwargs: Arguments that will be passed onto the plot function
+    :param ax_properties: A dict of axes properties. E.g. {'yscale': 'log'}
+    :param window_title: A string denoting name of the GUI window which holds the graph
+    :param tiled: An optional flag controlling whether to do a tiled or overlayed plot
+    :param waterfall: An optional flag controlling whether or not to do a waterfall plot
+    :return: The figure containing the plots
+    """
+    if plot_kwargs is None:
+        plot_kwargs = {}
+    _validate_plot_inputs(workspaces, spectrum_nums, wksp_indices, tiled, overplot)
+    workspaces = [ws for ws in workspaces if isinstance(ws, MatrixWorkspace)]
+
+    if spectrum_nums is not None:
+        kw, nums = 'specNum', spectrum_nums
     else:
-        axis = MantidAxType(kwargs.pop("axis", MantidAxType.SPECTRUM))
-        normalize_by_bin_width, kwargs = get_normalize_by_bin_width(
-            workspace, axes, **kwargs)
-        workspace_index, distribution, kwargs = get_wksp_index_dist_and_label(workspace, axis, **kwargs)
-        if axis == MantidAxType.BIN:
-            # Overwrite any user specified xlabel
-            axes.set_xlabel("Spectrum")
-            x, y, dy, dx = get_bins(workspace, workspace_index, with_dy)
-        elif axis == MantidAxType.SPECTRUM:
-            x, y, dy, dx = get_spectrum(workspace, workspace_index,
-                                        normalize_by_bin_width, with_dy, with_dx)
+        kw, nums = 'wkspIndex', wksp_indices
+
+    _add_default_plot_kwargs_from_settings(plot_kwargs, errors)
+
+    num_axes = len(workspaces) * len(nums) if tiled else 1
+
+    fig, axes = get_plot_fig(overplot, ax_properties, window_title, num_axes, fig)
+
+    # Convert to a MantidAxes if it isn't already. Ignore legend since
+    # a new one will be drawn later
+    axes = [MantidAxes.from_mpl_axes(ax, ignore_artists=[Legend]) if not isinstance(ax, MantidAxes) else ax
+            for ax in axes]
+
+    if tiled:
+        ws_index = [(ws, index) for ws in workspaces for index in nums]
+        for index, ax in enumerate(axes):
+            if index < len(ws_index):
+                _do_single_plot(ax, [ws_index[index][0]], errors, False, [ws_index[index][1]], kw, plot_kwargs)
+            else:
+                ax.axis('off')
+    else:
+        show_title = ("on" == ConfigService.getString("plots.ShowTitle").lower()) and not overplot
+        ax = overplot if isinstance(overplot, MantidAxes) else axes[0]
+        ax.axis('on')
+        _do_single_plot(ax, workspaces, errors, show_title, nums, kw, plot_kwargs)
+
+    # Can't have a waterfall plot with only one line.
+    if len(nums) == 1 and waterfall:
+        waterfall = False
+
+    # The plot's initial xlim and ylim are used to offset each curve in a waterfall plot.
+    # Need to do this whether the current curve is a waterfall plot or not because it may be converted later.
+    if not overplot:
+        datafunctions.set_initial_dimensions(ax)
+
+    if waterfall:
+        ax.set_waterfall(True)
+
+    if not overplot:
+        fig.canvas.set_window_title(figure_title(workspaces, fig.number))
+    else:
+        if ax.is_waterfall():
+            for i in range(len(nums)):
+                errorbar_cap_lines = datafunctions.remove_and_return_errorbar_cap_lines(ax)
+                datafunctions.convert_single_line_to_waterfall(ax, len(ax.get_lines()) - (i + 1))
+
+                if ax.waterfall_has_fill():
+                    datafunctions.waterfall_update_fill(ax)
+
+                ax.lines += errorbar_cap_lines
+
+    # This updates the toolbar so the home button now takes you back to this point.
+    # The try catch is in case the manager does not have a toolbar attached.
+    try:
+        fig.canvas.manager.toolbar.update()
+    except AttributeError:
+        pass
+
+    fig.canvas.draw()
+    # This displays the figure, but only works if a manager is attached to the figure.
+    # The try catch is in case a figure manager is not present
+    try:
+        fig.show()
+    except AttributeError:
+        pass
+
+    return fig
+
+
+def create_subplots(nplots, fig=None):
+    """
+    Create a set of subplots suitable for a given number of plots. A stripped down
+    version of plt.subplots that can accept an existing figure instance.
+
+    :param nplots: The number of plots required
+    :param fig: An optional figure. It is cleared before plotting the new contents
+    :return: A 2-tuple of (fig, axes)
+    """
+    import matplotlib.pyplot as plt
+    square_side_len = int(math.ceil(math.sqrt(nplots)))
+    nrows, ncols = square_side_len, square_side_len
+    if square_side_len * square_side_len != nplots:
+        # not a square number - square_side_len x square_side_len
+        # will be large enough but we could end up with an empty
+        # row so chop that off
+        if nplots <= (nrows - 1) * ncols:
+            nrows -= 1
+
+    if fig is None:
+        fig = plt.figure()
+    else:
+        fig.clf()
+    # annoyling this repl
+    nplots = nrows * ncols
+    gs = GridSpec(nrows, ncols)
+    axes = np.empty(nplots, dtype=object)
+    ax0 = fig.add_subplot(gs[0, 0], projection=PROJECTION)
+    axes[0] = ax0
+    for i in range(1, nplots):
+        axes[i] = fig.add_subplot(gs[i // ncols, i % ncols],
+                                  projection=PROJECTION)
+    axes = axes.reshape(nrows, ncols)
+
+    return fig, axes, nrows, ncols
+
+
+def raise_if_not_sequence(value, seq_name, element_type=None):
+    """
+    Raise a ValueError if the given object is not a sequence
+
+    :param value: The value object to validate
+    :param seq_name: The variable name of the sequence for the error message
+    :param element_type: An optional type to provide to check that each element
+    is an instance of this type
+    :raises ValueError: if the conditions are not met
+    """
+    accepted_types = (list, tuple, range)
+    if type(value) not in accepted_types:
+        raise ValueError("{} should be a list or tuple, "
+                         "instead found '{}'".format(seq_name,
+                                                     value.__class__.__name__))
+    if element_type is not None:
+        def raise_if_not_type(x):
+            if not isinstance(x, element_type):
+                if element_type == MatrixWorkspace:
+                    # If the workspace is the wrong type, log the error and remove it from the list so that the other
+                    # workspaces can still be plotted.
+                    LOGGER.warning("{} has unexpected type '{}'".format(x, x.__class__.__name__))
+                else:
+                    raise ValueError("Unexpected type: '{}'".format(x.__class__.__name__))
+
+        # Map in Python3 is an iterator, so ValueError will not be raised unless the values are yielded.
+        # converting to a list forces yielding
+        list(map(raise_if_not_type, value))
+
+
+def get_plot_fig(overplot=None, ax_properties=None, window_title=None, axes_num=1, fig=None):
+    """
+    Create a blank figure and axes, with configurable properties.
+    :param overplot: If true then plotting on figure will plot over previous plotting. If an axis object the overplotting
+    will be done on the axis passed in
+    :param ax_properties: A dict of axes properties. E.g. {'yscale': 'log'} for log y-axis
+    :param window_title: A string denoting the name of the GUI window which holds the graph
+    :param axes_num: The number of axes to create on the figure
+    :param fig: An optional figure object
+    :return: Matplotlib fig and axes objects
+    """
+    import matplotlib.pyplot as plt
+    if fig and overplot:
+        fig = fig
+    elif fig:
+        fig, _, _, _ = create_subplots(axes_num, fig)
+    elif overplot:
+        fig = plt.gcf()
+    else:
+        fig, _, _, _ = create_subplots(axes_num)
+
+    if not ax_properties:
+        ax_properties = {}
+        if ConfigService.getString("plots.xAxesScale").lower() == 'log':
+            ax_properties['xscale'] = 'log'
         else:
-            raise ValueError("Axis {} is not a valid axis number.".format(axis))
-        indices = None
-    return x, y, dy, dx, indices, axis, kwargs
-
-
-# ========================================================
-# Plot functions
-# ========================================================
-
-def _plot_impl(axes, workspace, args, kwargs):
-    """
-    Compute data and labels for plot. Used by workspace
-    replacement handlers to recompute data. See plot for
-    argument details
-    """
-    if 'LogName' in kwargs:
-        (x, y, FullTime, LogName, units, kwargs) = get_sample_log(workspace, **kwargs)
-        axes.set_ylabel('{0} ({1})'.format(LogName, units))
-        axes.set_xlabel('Time (s)')
-        if FullTime:
-            axes.xaxis_date()
-            axes.xaxis.set_major_formatter(mdates.DateFormatter('%H:%M:%S\n%b-%d'))
-            axes.set_xlabel('Time')
-        kwargs['linestyle'] = 'steps-post'
-    else:
-        normalize_by_bin_width, kwargs = get_normalize_by_bin_width(
-            workspace, axes, **kwargs)
-        x, y, _, _, indices, axis, kwargs = _get_data_for_plot(axes, workspace, kwargs)
-        if kwargs.pop('update_axes_labels', True):
-            _setLabels1D(axes, workspace, indices,
-                         normalize_by_bin_width=normalize_by_bin_width, axis=axis)
-    kwargs.pop('normalize_by_bin_width', None)
-    return x, y, args, kwargs
-
-
-def plot(axes, workspace, *args, **kwargs):
-    """
-    Unpack mantid workspace and render it with matplotlib. ``args`` and
-    ``kwargs`` are passed to :py:meth:`matplotlib.axes.Axes.plot` after special
-    keyword arguments are removed. This will automatically label the
-    line according to the spectrum number unless specified otherwise.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param specNum:   spectrum number to plot if MatrixWorkspace
-    :param wkspIndex: workspace index to plot if MatrixWorkspace
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the workspace is a MatrixWorkspace histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param normalize_by_bin_width: ``None`` (default) ask the workspace. It can override
-                          the value from distribution. Is implemented so get_normalize_by_bin_width
-                          only need to be run once.
-    :param LogName:   if specified, it will plot the corresponding sample log. The x-axis
-                      of the plot is the time difference between the log time and the first
-                      value of the `proton_charge` log (if available) or the sample log's
-                      first time.
-    :param StartFromLog: False by default. If True the time difference will be from the sample log's
-                      first time, even if `proton_charge` log is available.
-    :param FullTime:  False by default. If true, the full date and time will be plotted on the axis
-                      instead of the time difference
-    :param ExperimentInfo: for MD Workspaces with multiple :class:`mantid.api.ExperimentInfo` is the
-                           ExperimentInfo object from which to extract the log. It's 0 by default
-    :param axis: Specify which axis will be plotted. Use axis=MantidAxType.BIN to plot a bin,
-                 and axis=MantidAxType.SPECTRUM to plot a spectrum.
-                 The default value is axis=1, plotting spectra by default.
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimension to plot. *e.g.* to select the second axis to plot from a
-                    3D volume use ``indices=(5, slice(None), 10)`` where the 5/10 are the bins selected
-                    for the other 2 axes.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the second
-                       axis to plot from a 3D volume use ``slicepoint=(1.0, None, 2.0)`` where the 1.0/2.0 are
-                       the dimension selected for the other 2 axes.
-
-    For matrix workspaces with more than one spectra, either ``specNum`` or ``wkspIndex``
-    needs to be specified. Giving both will generate a :class:`RuntimeError`. There is no similar
-    keyword for MDHistoWorkspaces. These type of workspaces have to have exactly one non integrated
-    dimension
-    """
-    x, y, args, kwargs = _plot_impl(axes, workspace, args, kwargs)
-    return axes.plot(x, y, *args, **kwargs)
-
-
-def errorbar(axes, workspace, *args, **kwargs):
-    """
-    Unpack mantid workspace and render it with matplotlib. ``args`` and
-    ``kwargs`` are passed to :py:meth:`matplotlib.axes.Axes.errorbar` after special
-    keyword arguments are removed. This will automatically label the
-    line according to the spectrum number unless specified otherwise.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param specNum:   spectrum number to plot if MatrixWorkspace
-    :param wkspIndex: workspace index to plot if MatrixWorkspace
-    :param distribution: ``None`` (default) asks the global setting. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the workspace is a MatrixWorkspace histogram.
-    :param normalize_by_bin_width: Plot the workspace as a distribution. If None default to global
-                                   setting: config['graph1d.autodistribution']
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param axis: Specify which axis will be plotted. Use axis=MantidAxType.BIN to plot a bin,
-                  and axis=MantidAxType.SPECTRUM to plot a spectrum.
-                  The default value is axis=1, plotting spectra by default.
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimension to plot. *e.g.* to select the second axis to plot from a
-                    3D volume use ``indices=(5, slice(None), 10)`` where the 5/10 are the bins selected
-                    for the other 2 axes.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the second
-                       axis to plot from a 3D volume use ``slicepoint=(1.0, None, 2.0)`` where the 1.0/2.0 are
-                       the dimension selected for the other 2 axes.
-
-    For matrix workspaces with more than one spectra, either ``specNum`` or ``wkspIndex``
-    needs to be specified. Giving both will generate a :class:`RuntimeError`. There is no similar
-    keyword for MDHistoWorkspaces. These type of workspaces have to have exactly one non integrated
-    dimension
-    """
-    normalize_by_bin_width, kwargs = get_normalize_by_bin_width(
-        workspace, axes, **kwargs)
-    x, y, dy, dx, indices, axis, kwargs = _get_data_for_plot(
-        axes, workspace, kwargs, with_dy=True, with_dx=False)
-    if kwargs.pop('update_axes_labels', True):
-        _setLabels1D(axes, workspace, indices,
-                     normalize_by_bin_width=normalize_by_bin_width, axis=axis)
-
-    return axes.errorbar(x, y, dy, dx, *args, **kwargs)
-
-
-def scatter(axes, workspace, *args, **kwargs):
-    '''
-    Unpack mantid workspace and render it with matplotlib. ``args`` and
-    ``kwargs`` are passed to :py:meth:`matplotlib.axes.Axes.scatter` after special
-    keyword arguments are removed. This will automatically label the
-    line according to the spectrum number unless specified otherwise.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param specNum:   spectrum number to plot if MatrixWorkspace
-    :param wkspIndex: workspace index to plot if MatrixWorkspace
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the workspace is a MatrixWorkspace histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimension to plot. *e.g.* to select the second axis to plot from a
-                    3D volume use ``indices=(5, slice(None), 10)`` where the 5/10 are the bins selected
-                    for the other 2 axes.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the second
-                       axis to plot from a 3D volume use ``slicepoint=(1.0, None, 2.0)`` where the 1.0/2.0 are
-                       the dimension selected for the other 2 axes.
-
-    For matrix workspaces with more than one spectra, either ``specNum`` or ``wkspIndex``
-    needs to be specified. Giving both will generate a :class:`RuntimeError`. There is no similar
-    keyword for MDHistoWorkspaces. These type of workspaces have to have exactly one non integrated
-    dimension
-    '''
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        (x, y, _) = get_md_data1d(workspace, normalization, indices)
-        _setLabels1D(axes, workspace, indices)
-    else:
-        (wkspIndex, distribution, kwargs) = get_wksp_index_dist_and_label(workspace, **kwargs)
-        (x, y, _, _) = get_spectrum(workspace, wkspIndex, distribution)
-        _setLabels1D(axes, workspace)
-    return axes.scatter(x, y, *args, **kwargs)
-
-
-def contour(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`matplotlib.axes.Axes.contour`
-    but calculates the countour levels. Currently this only works with
-    workspaces that have a constant number of bins between spectra.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x, y, z = get_md_data2d_bin_centers(workspace, normalization, indices, transpose)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=False, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose)
-    return axes.contour(x, y, z, *args, **kwargs)
-
-
-def contourf(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`matplotlib.axes.Axes.contourf`
-    but calculates the countour levels. Currently this only works with
-    workspaces that have a constant number of bins between spectra.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x, y, z = get_md_data2d_bin_centers(workspace, normalization, indices, transpose)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=False, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose)
-    return axes.contourf(x, y, z, *args, **kwargs)
-
-
-def _pcolorpieces(axes, workspace, distribution, *args, **kwargs):
-    '''
-    Helper function for pcolor, pcolorfast, and pcolormesh that will
-    plot a 2d representation of each spectra. The polycollections or meshes
-    will be normalized to the same intensity limits.
-    :param axes: :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param pcolortype: this keyword allows the plotting to be one of pcolormesh or
-        pcolorfast if there is "mesh" or "fast" in the value of the keyword, or
-        pcolor by default
-    :return: A list of the pcolor pieces created
-    '''
-    (x, y, z) = get_uneven_data(workspace, distribution)
-    mini = numpy.min([numpy.min(i) for i in z])
-    maxi = numpy.max([numpy.max(i) for i in z])
-    if 'vmin' in kwargs:
-        mini = kwargs['vmin']
-    if 'vmax' in kwargs:
-        maxi = kwargs['vmax']
-    if 'norm' not in kwargs:
-        kwargs['norm'] = matplotlib.colors.Normalize(vmin=mini, vmax=maxi)
-    else:
-        if kwargs['norm'].vmin is None:
-            kwargs['norm'].vmin = mini
-        if kwargs['norm'].vmax is None:
-            kwargs['norm'].vmax = maxi
-
-    # setup the particular pcolor to use
-    pcolortype = kwargs.pop('pcolortype', '').lower()
-    if 'mesh' in pcolortype:
-        pcolor = axes.pcolormesh
-    elif 'fast' in pcolortype:
-        pcolor = axes.pcolorfast
-    else:
-        pcolor = axes.pcolor
-
-    pieces = []
-    for xi, yi, zi in zip(x, y, z):
-        XX, YY = numpy.meshgrid(xi, yi, indexing='ij')
-        pieces.append(pcolor(XX, YY, zi.reshape(-1, 1), **kwargs))
-
-    return pieces
-
-
-def pcolor(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`matplotlib.axes.Axes.pcolor`
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param axisaligned: ``False`` (default). If ``True``, or if the workspace has a variable
-                        number of bins, the polygons will be aligned with the axes
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x, y, z = get_md_data2d_bin_bounds(workspace, normalization, indices, transpose)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (aligned, kwargs) = check_resample_to_regular_grid(workspace, **kwargs)
-        (normalize_by_bin_width, kwargs) = get_normalize_by_bin_width(workspace, axes, **kwargs)
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        if aligned:
-            kwargs['pcolortype'] = ''
-            return _pcolorpieces(axes, workspace, distribution, *args, **kwargs)
+            ax_properties['xscale'] = 'linear'
+        if ConfigService.getString("plots.yAxesScale").lower() == 'log':
+            ax_properties['yscale'] = 'log'
         else:
-            (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=True, transpose=transpose)
-            _setLabels2D(axes, workspace, transpose=transpose, normalize_by_bin_width=normalize_by_bin_width)
-    return axes.pcolor(x, y, z, *args, **kwargs)
+            ax_properties['yscale'] = 'linear'
+    if ax_properties:
+        for axis in fig.axes:
+            axis.set(**ax_properties)
+    if window_title:
+        fig.canvas.set_window_title(window_title)
+
+    return fig, fig.axes
 
 
-def pcolorfast(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`matplotlib.axes.Axes.pcolorfast`
+# -----------------------------------------------------------------------------
+# Pricate Methods
+# -----------------------------------------------------------------------------
+def _validate_plot_inputs(workspaces, spectrum_nums, wksp_indices, tiled=False, overplot=False):
+    """Raises a ValueError if any arguments have the incorrect types"""
+    if spectrum_nums is not None and wksp_indices is not None:
+        raise ValueError("Both spectrum_nums and wksp_indices supplied. "
+                         "Please supply only 1.")
 
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param axisaligned: ``False`` (default). If ``True``, or if the workspace has a variable
-                        number of bins, the polygons will be aligned with the axes
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x, y, z = get_md_data2d_bin_bounds(workspace, normalization, indices, transpose)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (aligned, kwargs) = check_resample_to_regular_grid(workspace, **kwargs)
-        (normalize_by_bin_width, kwargs) = get_normalize_by_bin_width(workspace, axes, **kwargs)
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        if aligned:
-            kwargs['pcolortype'] = 'fast'
-            return _pcolorpieces(axes, workspace, distribution, *args, **kwargs)
-        else:
-            (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=True, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose, normalize_by_bin_width=normalize_by_bin_width)
-    return axes.pcolorfast(x, y, z, *args, **kwargs)
+    if tiled and overplot:
+        raise ValueError("Both tiled and overplot flags set to true. "
+                         "Please set only one to true.")
+
+    raise_if_not_sequence(workspaces, 'workspaces', MatrixWorkspace)
+
+    if spectrum_nums is not None:
+        raise_if_not_sequence(spectrum_nums, 'spectrum_nums')
+
+    if wksp_indices is not None:
+        raise_if_not_sequence(wksp_indices, 'wksp_indices')
 
 
-def pcolormesh(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`matplotlib.axes.Axes.pcolormesh`.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param axisaligned: ``False`` (default). If ``True``, or if the workspace has a variable
-                        number of bins, the polygons will be aligned with the axes
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x, y, z = get_md_data2d_bin_bounds(workspace, normalization, indices, transpose)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (aligned, kwargs) = check_resample_to_regular_grid(workspace, **kwargs)
-        (normalize_by_bin_width, kwargs) = get_normalize_by_bin_width(workspace, axes, **kwargs)
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        if aligned:
-            kwargs['pcolortype'] = 'mesh'
-            return _pcolorpieces(axes, workspace, distribution, *args, **kwargs)
-        else:
-            (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=True, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose, normalize_by_bin_width=normalize_by_bin_width)
-    return axes.pcolormesh(x, y, z, *args, **kwargs)
+def _add_default_plot_kwargs_from_settings(plot_kwargs, errors):
+    if 'linestyle' not in plot_kwargs:
+        plot_kwargs['linestyle'] = ConfigService.getString("plots.line.Style")
+    if 'linewidth' not in plot_kwargs:
+        plot_kwargs['linewidth'] = float(ConfigService.getString("plots.line.Width"))
+    if 'marker' not in plot_kwargs:
+        plot_kwargs['marker'] = MARKER_MAP[ConfigService.getString("plots.marker.Style")]
+    if 'markersize' not in plot_kwargs:
+        plot_kwargs['markersize'] = float(ConfigService.getString("plots.marker.Size"))
+    if errors:
+        if 'capsize' not in plot_kwargs:
+            plot_kwargs['capsize'] = float(ConfigService.getString("plots.errorbar.Capsize"))
+        if 'capthick' not in plot_kwargs:
+            plot_kwargs['capthick'] = float(ConfigService.getString("plots.errorbar.CapThickness"))
+        if 'errorevery' not in plot_kwargs:
+            plot_kwargs['errorevery'] = int(ConfigService.getString("plots.errorbar.errorEvery"))
+        if 'elinewidth' not in plot_kwargs:
+            plot_kwargs['elinewidth'] = float(ConfigService.getString("plots.errorbar.Width"))
 
 
-def imshow(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`matplotlib.axes.Axes.imshow`.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param axisaligned: ``False`` (default). If ``True``, or if the workspace has a variable
-                        number of bins, the polygons will be aligned with the axes
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x, y, z = get_md_data2d_bin_bounds(workspace, normalization, indices, transpose)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (aligned, kwargs) = check_resample_to_regular_grid(workspace, **kwargs)
-        (normalize_by_bin_width, kwargs) = get_normalize_by_bin_width(workspace, axes, **kwargs)
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        if aligned:
-            (x, y, z) = get_matrix_2d_ragged(workspace, normalize_by_bin_width, histogram2D=True, transpose=transpose)
-        else:
-            (x, y, z) = get_matrix_2d_data(workspace, distribution=distribution, histogram2D=True, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose, normalize_by_bin_width=normalize_by_bin_width)
-    if 'extent' not in kwargs:
-        if x.ndim == 2 and y.ndim == 2:
-            kwargs['extent'] = [x[0, 0], x[0, -1], y[0, 0], y[-1, 0]]
-        else:
-            kwargs['extent'] = [x[0], x[-1], y[0], y[-1]]
-    return mantid.plots.modest_image.imshow(axes, z, *args, **kwargs)
-
-
-def tripcolor(axes, workspace, *args, **kwargs):
-    '''
-    To be used with non-uniform grids. Currently this only works with workspaces
-    that have a constant number of bins between spectra or with
-    MDHistoWorkspaces.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-
-    See :meth:`matplotlib.axes.Axes.tripcolor` for more information.
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x_temp, y_temp, z = get_md_data2d_bin_centers(workspace, normalization, indices, transpose)
-        x, y = numpy.meshgrid(x_temp, y_temp)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=False, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose)
-    return axes.tripcolor(x.ravel(), y.ravel(), z.ravel(), *args, **kwargs)
-
-
-def tricontour(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`mantid.plots.contour`, but works
-    for non-uniform grids. Currently this only works with workspaces
-    that have a constant number of bins between spectra or with
-    MDHistoWorkspaces.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-
-    See :meth:`matplotlib.axes.Axes.tricontour` for more information.
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x_temp, y_temp, z = get_md_data2d_bin_centers(workspace, normalization, indices, transpose)
-        x, y = numpy.meshgrid(x_temp, y_temp)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=False, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose)
-    # tricontour segfaults if many z values are not finite
-    # https://github.com/matplotlib/matplotlib/issues/10167
-    x = x.ravel()
-    y = y.ravel()
-    z = z.ravel()
-    condition = numpy.isfinite(z)
-    x = x[condition]
-    y = y[condition]
-    z = z[condition]
-    return axes.tricontour(x, y, z, *args, **kwargs)
-
-
-def tricontourf(axes, workspace, *args, **kwargs):
-    '''
-    Essentially the same as :meth:`mantid.plots.contourf`, but works
-    for non-uniform grids. Currently this only works with workspaces
-    that have a constant number of bins between spectra or with
-    MDHistoWorkspaces.
-
-    :param axes:      :class:`matplotlib.axes.Axes` object that will do the plotting
-    :param workspace: :class:`mantid.api.MatrixWorkspace` or :class:`mantid.api.IMDHistoWorkspace`
-                      to extract the data from
-    :param distribution: ``None`` (default) asks the workspace. ``False`` means
-                         divide by bin width. ``True`` means do not divide by bin width.
-                         Applies only when the the matrix workspace is a histogram.
-    :param normalization: ``None`` (default) ask the workspace. Applies to MDHisto workspaces. It can override
-                          the value from displayNormalizationHisto. It checks only if
-                          the normalization is mantid.api.MDNormalization.NumEventsNormalization
-    :param indices: Specify which slice of an MDHistoWorkspace to use when plotting. Needs to be a tuple
-                    and will be interpreted as a list of indices. You need to use ``slice(None)`` to
-                    select which dimensions to plot. *e.g.* to select the last two axes to plot from a
-                    3D volume use ``indices=(5, slice(None), slice(None))`` where the 5 is the bin selected
-                    for the first axis.
-    :param slicepoint: Specify which slice of an MDHistoWorkspace to use when plotting in the dimension units.
-                       You need to use ``None`` to select which dimension to plot. *e.g.* to select the last
-                       two axes to plot from a 3D volume use ``slicepoint=(1.0, None, None)`` where the 1.0 is
-                       the value of the dimension selected for the first axis.
-    :param transpose: ``bool`` to transpose the x and y axes of the plotted dimensions of an MDHistoWorkspace
-
-    See :meth:`matplotlib.axes.Axes.tricontourf` for more information.
-    '''
-    transpose = kwargs.pop('transpose', False)
-    if isinstance(workspace, mantid.dataobjects.MDHistoWorkspace):
-        (normalization, kwargs) = get_normalization(workspace, **kwargs)
-        indices, kwargs = get_indices(workspace, **kwargs)
-        x_temp, y_temp, z = get_md_data2d_bin_centers(workspace, normalization, indices, transpose)
-        x, y = numpy.meshgrid(x_temp, y_temp)
-        _setLabels2D(axes, workspace, indices, transpose)
-    else:
-        (distribution, kwargs) = get_distribution(workspace, **kwargs)
-        (x, y, z) = get_matrix_2d_data(workspace, distribution, histogram2D=False, transpose=transpose)
-        _setLabels2D(axes, workspace, transpose=transpose)
-    # tricontourf segfaults if many z values are not finite
-    # https://github.com/matplotlib/matplotlib/issues/10167
-    x = x.ravel()
-    y = y.ravel()
-    z = z.ravel()
-    condition = numpy.isfinite(z)
-    x = x[condition]
-    y = y[condition]
-    z = z[condition]
-    return axes.tricontourf(x, y, z, *args, **kwargs)
-
-
-def update_colorplot_datalimits(axes, mappables):
+def _validate_workspace_names(workspaces):
     """
-    For an colorplot (imshow, pcolor*) plots update the data limits on the axes
-    to circumvent bugs in matplotlib
-    :param mappables: An iterable of mappable for this axes
+    Checks if the workspaces passed into a plotting function are workspace names, and
+    retrieves the workspaces if they are.
+    This function assumes that we do not have a mix of workspaces and workspace names.
+    :param workspaces: A list of workspaces or workspace names
+    :return: A list of workspaces
     """
-    # ax.relim in matplotlib < 2.2 doesn't take into account of images
-    # and it doesn't support collections at all as of verison 3 so we'll take
-    # over
-    if not isinstance(mappables, collections.Iterable):
-        mappables = [mappables]
-    xmin_all, xmax_all, ymin_all, ymax_all = _LARGEST, _SMALLEST, _LARGEST, _SMALLEST
-    for mappable in mappables:
-        xmin, xmax, ymin, ymax = get_colorplot_extents(mappable)
-        xmin_all, xmax_all = min(xmin_all, xmin), max(xmax_all, xmax)
-        ymin_all, ymax_all = min(ymin_all, ymin), max(ymax_all, ymax)
-    axes.update_datalim(((xmin_all, ymin_all), (xmax_all, ymax_all)))
-    axes.autoscale()
-
-
-def get_colorplot_extents(mappable):
-    """
-    Return the extent of the given mappable
-    :param mappable: A 2D mappable object
-    :return: (left, right, bottom, top)
-    """
-    if isinstance(mappable, mimage.AxesImage):
-        xmin, xmax, ymin, ymax = mappable.get_extent()
-    elif isinstance(mappable, mcoll.QuadMesh):
-        # coordinates are vertices of the grid
-        coords = mappable._coordinates
-        xmin, ymin = coords[0][0]
-        xmax, ymax = coords[-1][-1]
-    elif isinstance(mappable, mcoll.PolyCollection):
-        xmin, ymin = mappable._paths[0].get_extents().min
-        xmax, ymax = mappable._paths[-1].get_extents().max
+    try:
+        raise_if_not_sequence(workspaces, 'workspaces', string_types)
+    except ValueError:
+        return workspaces
     else:
-        raise ValueError("Unknown mappable type '{}'".format(type(mappable)))
+        return AnalysisDataService.Instance().retrieveWorkspaces(workspaces, unrollGroups=True)
 
-    return xmin, xmax, ymin, ymax
+
+def _do_single_plot(ax, workspaces, errors, set_title, nums, kw, plot_kwargs):
+    # do the plotting
+    plot_fn = ax.errorbar if errors else ax.plot
+    for ws in workspaces:
+        for num in nums:
+            plot_kwargs[kw] = num
+            plot_fn(ws, **plot_kwargs)
+    ax.make_legend()
+    if set_title:
+        title = workspaces[0].name()
+        ax.set_title(title)
+
+
