+++ qt/applications/workbench/workbench/plotting/figureerrorsmanager.py
@@ -11,8 +11,8 @@ Controls the dynamic displaying of errors for line on the plot
 from matplotlib.container import ErrorbarContainer
 from matplotlib.lines import Line2D
 
-from mantid.plots import helperfunctions, MantidAxes
-from mantid.plots.helperfunctions import get_data_from_errorbar_container, set_errorbars_hidden
+from mantid.plots import datafunctions, MantidAxes
+from mantid.plots.datafunctions import get_data_from_errorbar_container, set_errorbars_hidden
 from mantid.plots.legend import LegendProperties
 from mantidqt.widgets.plotconfigdialog.curvestabwidget import curve_has_errors, CurveProperties, remove_curve_from_ax
 
@@ -51,14 +51,14 @@ class FigureErrorsManager(object):
         new_curve = cls.replot_curve(ax, curve, plot_kwargs)
 
         if isinstance(ax, MantidAxes):
-            errorbar_cap_lines = helperfunctions.remove_and_return_errorbar_cap_lines(ax)
+            errorbar_cap_lines = datafunctions.remove_and_return_errorbar_cap_lines(ax)
         else:
             errorbar_cap_lines = []
 
         ax.lines.insert(curve_index, ax.lines.pop())
 
         if isinstance(ax, MantidAxes) and ax.is_waterfall():
-            helperfunctions.convert_single_line_to_waterfall(ax, curve_index)
+            datafunctions.convert_single_line_to_waterfall(ax, curve_index)
 
         ax.lines += errorbar_cap_lines
