+++ qt/applications/workbench/workbench/plotting/propertiesdialog.py
@@ -12,7 +12,7 @@ from __future__ import (absolute_import, unicode_literals)
 # std imports
 
 # 3rdparty imports
-from mantid.plots.helperfunctions import update_colorbar_scale
+from mantid.plots.datafunctions import update_colorbar_scale
 from mantidqt.plotting.figuretype import FigureType, figure_type
 from mantidqt.utils.qt import load_ui
 from matplotlib.collections import QuadMesh
