+++ qt/python/mantidqt/widgets/sliceviewer/view.py
@@ -14,7 +14,7 @@ from matplotlib.figure import Figure
 from matplotlib.transforms import Bbox, BboxTransform
 
 import mantid.api
-from mantid.plots.helperfunctions import get_normalize_by_bin_width
+from mantid.plots.datafunctions import get_normalize_by_bin_width
 from mantidqt.MPLwidgets import FigureCanvas
 from mantidqt.widgets.colorbar.colorbar import ColorbarWidget
 from .dimensionwidget import DimensionWidget
